Seperate file for config routines
